Redirect aplicada a login (profesor, alumno), favicon, js. mensaje eliminar

+++ src/main/webapp/privado/profesor.jsp
@@ -10,30 +10,31 @@
 
 <h1>Bienvenido PROFESOR</h1>
 
-${cursosAlumnos }
 
 <p>${mensaje}</p>
 
 <h2>Mis Cursos</h2>
 	<div>
-		<table class="table table table-sm table-bordered table-info">
+		<table class="table table-sm table-bordered table-info">
 		  <thead>
 		    <tr>
 		      <th scope="col">id</th>
 		      <th scope="col">Curso</th>
 		      <th scope="col">Identificador</th>
 		      <th scope="col">Horas</th>
+		       <th scope="col">Alumnos Inscritos</th>
 		      <th scope="col">Operacion</th>
 		    </tr>
 		  </thead>
 		  <tbody>
-		     <c:forEach items="${cursos}" var="curso">
+		     <c:forEach items="${cursosAlumnos}" var="curso">
 		    	<tr class="table-light">
 			      <th scope="row">${curso.id}</th>
 			      <td>${curso.nombre}</td>
 			      <td>${curso.identificador}</td>
 			      <td>${curso.horas}</td>
-			      <td class="blockquote text-center"><a href="eliminar-curso?id=${curso.id}"><i class="fa fa-trash"></i></a></td>
+			      <td>${curso.numAlumnos}</td>
+			      <td class="blockquote text-center"><a href="eliminar-curso?id=${curso.id}" onclick="confirmar('${curso.nombre}')"><i class="fa fa-trash"></i></a></td>
 		    	</tr>
 		    </c:forEach>
 		  </tbody>
@@ -53,18 +54,18 @@ ${cursosAlumnos }
 			
 				<div class="form-group">	
 					<label for="nombre">Nombre del Curso </label>
-					<input 	type="text" class="form-control"  name="curso" id="curso" value=" " 
+					<input 	type="text" class="form-control"  name="curso" id="curso" value="" 
 										placeholder="Debe tener entre 3 y 100 caracteres" autofocus required>
 				</div>	
 				<div class="form-group">	
 					<label for="nombre">Identificador </label>
-					<input 	type="text" class="form-control"  name="identificador" id="identificador" value=" " 
+					<input 	type="text" class="form-control"  name="identificador" id="identificador" value="" 
 										placeholder="Debe tener entre 3 y 100 caracteres" autofocus required>
 				</div>	
 				<div class="form-group">	
 					<label for="nombre">Horas del Curso </label>
-					<input 	type="text" class="form-control"  name="horas" id="horas" value=" " 
-										placeholder="Debe tener entre 3 y 100 caracteres" autofocus required>
+					<input 	type="text" class="form-control"  name="horas" id="horas" value="" 
+										placeholder="Debe tener entre 2 y 100 caracteres" autofocus required>
 				</div>	
 				
 				<div class="form-group">
